+++ primitives/executor/src/lib.rs
@@ -22,10 +22,17 @@
 #[cfg(not(feature = "std"))]
 extern crate alloc;
 
+// use std::collections::HashSet;
+// use std::sync::Arc;
+
+// use exec_membership_runtime::ExecutorMemberApi;
 #[cfg(feature = "std")]
 use serde::Serialize;
 
 use codec::{Encode, Decode, Input, Codec};
+// use sp_api::{ApiErrorFor, BlockId, BlockT, ProvideRuntimeApi};
+// use sp_blockchain::HeaderBackend;
+// use sp_keystore::CryptoStore;
 use sp_runtime::{ConsensusEngineId, RuntimeDebug, traits::NumberFor};
 use sp_std::borrow::Cow;
 use sp_std::vec::Vec;